Adds option for Mac builds to be done inside or outside app bundles.
(Defaults to inside app bundle;  the old behaviour)

If you compile with the VS_APPBUNDLE option turned off, we'll expect the
data files to be in a 'Data' directory next to the executable, the same
as on other platforms.  This is actually surprisingly more convenient
during development, since OS X's Gatekeeper system is less likely to get
annoyed at you over app signing if your program hasn't been moved into
an app bundle.