Adjust CPU frame rate limiter to limit to *double* monitor refresh rate.
I'm not tremendously happy with our frame rate limiter, but I've bumped
it to limit us to 2x the monitor's refresh rate, which means that we're
less likely to hit the limiter while the driver is secretly setting up
triple-buffering or due to small timer inaccuracies.

I'm going to have to revisit this someday, but I'm generally happier
with this approach than what we had before.