Improve vsFile::MODE_ReadCompressed_Progressive.
Due to an apparently not-entirely-completed code refactoring, files
loaded in "ReadCompressed_Progressive" mode were being completely
decompressed twice;  once during the initial open, and then again as
data is actually being read.

Ostensibly we might have been doing it so that we could tell how big the
file was going to be uncompressed, but we weren't actually setting that
size onto the file's "m_length" variable and we don't allow seeking
around inside of a progressively compressed file (for hopefully obvious
reasons), so.. it was kind of a big expensive no-op.  So I've removed
it!

Now we only decompress the file once, on demand, instead of twice.