vsFile::MODE_ReadCompressed_Progressive now loads *properly* progressively
Previously, we loaded the whole compressed file into memory and
decompressed chunks of it into memory one at a time, as games read data
from the file.

Now we handle both parts progressively;  we load chunks of compressed
data from disk as we need them, and we also decompress that data as it's
needed by the load process, so now we never have *either* the full
compressed *or* the full decompressed save file in memory all at once;
instead we hold up to 100KB of each in memory at any given time.

There's still some work that could be done to make this process a little
smarter, so that we fetch data from disk in larger chunks only when
absolutely required.  But as a first step, this wins up some big memory
savings.

In MT2, this can save a few dozen megabytes of memory during the
load of a large save.