Even smarter progressive loading of compressed file data.
Data now passes through two 100KB buffers.

Disk File -> 100KB Compressed Buffer -> 100KB Decompressed Buffer -> Load

When loading code requests data from the file, we deliver data from the
decompressed buffer until it has no decompressed data left to offer.
When that happens, we decompress as much data from the compressed buffer
as we can fit into the decompressed buffer and repeat.

If we still can't provide enough decompressed data by decompressing the
data in the Compressed Buffer, then we fill up our Compressed Buffer with
more compressed data from disk and then decompress it into the
decompressed buffer.  And we repeat this process until either we have
enough data in the decompressed buffer to fulfill the read request, or
until the file has no more data to load and we tell the caller how much
data we actually have.

As a result, we typically go to disk for about 100KB reads, and
decompress data around 100KB at a time.

Further improvement to consider:

Right now we require the decompressed buffer to simultaneously contain
all the data asked for by the caller, which has a couple nasty edge
cases.  For example, as our buffer is 100KB in size, if a client asked
us to give them 101KB of data in a single 'Read' action, that request
would fail because we don't actually have the capacity to hold that much
data at once.

Either we should be copying data out into the buffer provided by the
caller each time around the loop (so we don't actually have to store all
the data simultaneously inside our work buffers), or maybe we don't need
the decompressed buffer at all;  we just keep the compressed buffer and
decompress it a few bytes at a time into individual read requests.

Not sure what the performance implications are (or aren't) in invoking
zlib with hundreds of tiny inflate() requests vs. just one big one.
For the moment, I'm going with the "just one big one" because it feels
like it's less likely to have performance complications!