Adds a different codepath for vsFile::Rename for OSX.
This uses POSIX ::rename instead of std::filesystem::rename.

std::filesystem::rename was only implemented in OSX in OSX 10.15, which
is only about two years old;  that's *much* too recent for my tastes!
So I'm yanking that back out (for OSX builds only) and falling back to
the standard POSIX ::rename, yay!

This means that we can drop back to our old 10.9 deployment settings for
Mac builds, which is 9 years old, which is a period I feel a *lot* more
comfortable requiring of players!