Call fflush() after writing to log files.
In theory this shouldn't block;  it just pushes data from a buffer in
the game RAM to a buffer in kernel RAM, from which is gets written out
to disk.  Since it's in the kernel, it can still write to disk
asynchronously from the game's point of view so it shouldn't affect our
frame rate or etc.

This fixes a problem of the game log being truncated if the game
outright crashes (which thankfully isn't very common these days!)