Moves a bunch of math from the vsRenderQueue to vsCamera.
We now have utility functions on the vsCamera and vsCamera3D for
calculating the final WorldToViewMatrix for a camera, (including fixes
for screen orientation), instead of putting all of that math into the
vsRenderQueue.

This simplifies the code a *lot*, and also makes the math accessible to
any game code that wants to make use of it.

Also in this commit, we moved to a built-in shared vsRenderQueue across
all scenes (assumption:  we're never drawing two or more scenes
simultaneously on different threads), and also made a lot more vsCamera
member functions const-correct.

Also also in this commit, vsRenderQueues no longer take any parameters
in the constructor;  instead, they always create a 50kb generic display
list which is dynamically resizable.  This commit also removes the
"stageCount" parameter which hasn't been being used for years.