vsFragment::SetSimple() can now indicate buffer ownership state.
Previously, vsFragment::SetSimple() just automatically took over
ownership of any buffers passed into it.  This is still the default, but
now you can pass in an enumerated value to specify whether VBO and/or
IBO should be considered 'owned' by the fragment and therefore be
destroyed if/when the fragment is destroyed.