vsInput now uses CursorPos struct to store text cursor positions.
vsInput's built-in text processing only handles text as a 1D string, but
if you're using an external CursorHandler you've now got a spot to
specify which line the cursor is on, and vsInput will (mostly!) try to
maintain it.

Also in this commit:

vsStringModeCursorHandler renamed to vsInput::CursorHandler.

CursorHandler now receives 'TextInput' messages and should handle them
itself.  CursorHandler is expected to return *both* the updated string
*and* the new cursor position.