Fix vsSerialiserRead when trying to read data bigger than its buffer
The vsSerialiserRead maintains an internal 1kb buffer which it services
read requests from.  If it was asked to read a data blob that's larger
than that 1kb buffer in a single read, it would generate a spurious "the
file is truncated" error message and assert out.

Now, if the vsSerialiserRead is asked to read a larger data blob, it
resizes its internal buffer so that it can actually do that, and the
only reason to get a "the file is truncated" error message would be if
the file was *actually* truncated.

Additionally, made the default buffer size 2kb instead of 1kb, so it
should happen less often anyway.