Changes to how we handle texture rendering settings
Previously, we had a "clampU" and "clampV" setting on the
vsMaterialInternal.  This is no longer the case!

It was awkward because it was set universally on the material, and then
*doubly* awkward because we actually applied those settings to the
texture object, so if multiple materials referred to the same texture
with differnet 'clamp' settings, the last-material-loaded version would
'win' and have its clamp settings used for all other materials.

It was *also* awkward because you had to have the same clamp settings
for all textures.

Now we instead have moved 'clamp' management (and also 'linear/nearest'
sampling mode management) over to the vsTexture object.

So when you create a vsTexture, you can set how you want its sampling
modes to work, and different vsTexture objects referring to the same
vsTextureInternal can use different modes and it all just works.

We store "current configuration" cacheing data on the vsTextureInternal
so that if twenty things want to same modes then we don't reset those
modes during each draw.

This should *mostly* be a pretty invisible change for users, except:

"clampU" and "clampV" lines in a material file no longer have any
effect.  Instead, these are modifiers on "texture" lines.  So if you
want a texture to be clamped, you define it like this:

`texture "filename.png" clampU clampV`

Or even easier:

`texture "filename.png" clampUV`

You can also specify `linear` or `nearest` sampling modes in the same
way ('linear' is the default)