Improve CPU-based Max-FPS limiter.
Previously we applied the delay at the *start* of a frame, now we apply
it at the *end* (after 'Present').

Previously, there was a one-frame delay on the FPS limiter, so if there
was a short frame we'd insert a delay at the start of the next frame,
which might end up running long such that our delay just made everything
worse.

Now the delay is imposed during the same frame.  This probably isn't a
*real* improvement since video drivers are good at doing tricks about
deciding where the 'beginning' and 'end' of frames are, but.. it looks
better this way in my profiler and that's the most important thing,
right?