Fixed an issue with non-uniform scaling in vsTransforms.
Additionally, added some new functions to vsMatrix4x4:

- Translate()

Modifies the 'w' component of the matrix, in order to move the
position being referred to.  This change happens in *world* space.

- Rotate()

Given a vsMatrix3x3 parameter, modifies the 'xyz' components of the
matrix to rotate its orientation by the specified additional rotation.

- Scale()

This function already existed. It 'scales' the matrix by changing the
lengths of the 'xyz' components.  There was a bug in this function
which applied the scale incorrectly in the case of non-uniform scales.
Now fixed!