Fixes to vsWeakPointer.
There was previously some unsafe C-style casting in our weak pointer
implementation, which could break in the case of multiple inheritance.

Our weak pointer implementation originally came from a friend of mine,
but this change breaks it enough that I've removed all remaining traces
of his code, including the header guards and commenting style (ew,
Javadoc), to make it sit more nicely in our codebase.

But mostly this is about fixing the multiple inheritance problem.  We
now actually store the typed pointer in a safe way, so no casting is
required at all when accessing the contents of the weak pointer.