Fixes dynamic batching of triangle strips.
Triangle strips, of course, swap winding order on every triangle in the
strip, and OpenGL undoes that.

We were naively ignoring that issue when creating dynamic batches from
triangle strips, which meant that any material which was trying to do
backface culling would have half its triangles culled from anything that
had been rendered via a triangle strip that was added to a dynamic batch.