Experimental switching from fmtlib to tinyformat.
fmtlib is mostly trying to provide a python-style interface to string
formatting, and has been slowly removing features from its C-style
printf-based API.  A combination of old releases of fmtlib breaking
under newer releases of clang and printf features going away have been
pulling me away from it.

TinyFormat is an alternate (also free+open source) implementation which
is smaller, simpler, and entirely based on the printf standard which
we're already using.  Swapping over to this was *heaps* easier than
trying to fix our issues with fmtlib!

I'm going to live with this for a couple days and see how I like it over
time, and may eventually merge it into our master branch if it goes well.