Add read/write locks around constructing the task thread table.
There's very limited chance for a conflict here, but it's theoretically
possible that two threads starting up could both be trying to insert
theiir data into the thread table simultaneously.  Now we protect
against that!

Also, we set things up so that vsLog messages trace out task names
(which are reliable and set by the game, instead of id numbers which are
potentially unreliable and different from run to run)