Fixes pipeline render target reuse after a window resize.
If a render pipeline requested a render target based upon the renderable
screen size, it was matching that request against existing render
targets which requested the same mipmap level, without considering
whether the whole window had changed size.

So for example, if you had a 640x480 window and your render pipeline was
using an offscreen window-sized buffer for an effect, if you resized the
window to 1024x768 and your pipeline again requested to use an offscreen
window-sized render target, the system would give you that same old
640x480 render target, seeing as your new request and the old one were
both asking for "a render target that matches the screen".

Now it's smarter and actually matches the number of pixels in that case
of screen-relative sizing!