Experimental change for VBO reuse.
Here instead of creating and destroying VBOs in the OpenGL context, we
keep track of the VBOs we've previously allocated and reuse them for
future buffers.

We orphan the buffers before putting them aside, so the video driver
shouldn't be hanging on to any data for those buffers that we would
otherwise be destroying.