Fixes a stack overflow when trying to format absurdly huge numbers.
Was a combination of bugs which eventually led to a stack overflow
because abs(-2147483648) == -2147483648.

(I mean, technically it doesn't equal that;  it's undefined behaviour to
try to get the absolute value of the largest possible negative int value
as it results in an integer wraparound, so technically the compiler
could legally do anything it wanted and you couldn't say it was
disobeying the specs, but most compilers will wrap around like this and
result in the same value you passed into 'abs'.  And in this case, the
function was calling itself with the abs'd value which resulted in an
infinite regression)