Adds input binding for control keys.
This makes it so you can bind a single input code to "control+z", for
example, and not have it trigger on simple "z" and always have to check
for 'control' yourself.

Downside is that we're *not* currently masking out NON-control.  That
is, a control that's mapped to Z *will* be triggered at the same time as
your control+Z mapping.

I need to make this easier to handle!  But that'll be a future update.