Adds CT_KeyboardKeycode for keycode-based keyboard controls.
There are two ways to handle keyboard input;  scancode or keycode.

Scancode refers to the physical position of a key on a keyboard (WASD,
for example, is commonly mapped by scancode), Keycode refers to what
letter is produced when you press the key.

As a general rule, when you're doing video game controls, you want to be
using scancodes to put keys in convenient locations.  But sometimes
you're picking keys based upon a mnemonic ('M' for 'Map', for example),
and so maybe you want to actually use the 'M' key no matter what
keyboard layout somebody's using.

The danger, of course, is always that you're going to end up with keys
overlapping if you use both scancodes and keycodes.

For example, it's common to use the 'Z' keycode to for 'Undo' features.
However, in an AZERTY layout, the 'Z' key is where 'W' is on a QWERTY
layout, and so your scancode WASD for movement is conflicting with your
keycode Z for 'undo'!

General rule of thumb is to only use scancodes or only use keycodes.
Here, we generally prefer scancodes.  But sometimes you do need keycodes
(for example, for 'undo'), and so this supports that use-case.  Just be
careful with it!