Move the debug timing visualisation (when enabled) into explicit pipelines.
The engine now draws the debug timing by putting it into a pipeline
explicitly.  When DEBUG_SCENE is enabled in the engine, it goes into the
engine's default pipeline.  If your game is providing its own rendering
pipeline, (as MMORPG Tycoon 2 does), then you'll need to add it to your
own pipeline manually, if you want it!