Starting to try to get vsFile to handle corrupt files a little more nicely
Here we're dealing with reading corrupted compressed files;  the file
and a streamed serialiser can now go into a "not okay" state which the
caller can check before further using the data.

This isn't enough, but..  it's maybe a first step toward more slightly
more robust handling of file errors?