Do not flag SIGTERM for crash reporting on Linux/Mac.
As per the documentation of standard termination signals:

"The SIGTERM is a generic signal used to cause program termination.
(...) It is the normal way to politely ask a program to terminate."

We no longer hook that signal and treat it as if the game was crashing;
we now instead allow SDL2 to catch it and treat it as a normal quit
request, letting us exit cleanly and *not* send a crash report on the
next run of the game.