Fixes VRAM tracking code during window resizing.
There was a bug which was preventing resizing of the main window from
updating our VRAM usage estimate.  This meant that if you started with a
small window, made it larger, and then shut down the program, the VRAM
usage estimate would throw an assertion as we were claiming we were
freeing more VRAM than we'd allocated.

We now handle those resizes correctly!