Cleanup to primitive reset handling.
Now primitive reset is something that gets set on individual index
buffers to state whether or not that particular index buffer uses
primitive reset.

Only implemented for triangle strips and triangle fans (because why
would you want to use primitive reset on a triangle list?), and only
really *tested* with triangle strips.  But it's an initial proof of
concept.