Fixes issues with texture mipmap generation.
We were stomping over texture binding, which could lead to subsequent
draws (usually involving the texture which had just had a mipmap
generated) ended up without having the texture bound during the draw.

This fix is ugly, it's basically exposing some global state variables so
we can return the texture that the *renderer* thinks is bound back, when
we're done building the mipmaps.