Set things up to make it easy to swap between packed normals and unpacked.
Now set back to *unpacked*.

We've had a couple of MT2 players who are stuck with old AMD drivers
(typically version 22.6.1), which fails with our packed normals for some
reason;  just returning 0 values.

For those who can update those old drivers to the latest (25.5.1), the
packed normals work great.  But... it's just not worth handling the
extra support requests for the mere 40 megabytes of VRAM we were saving.