Adds a new "preserve alpha" draw mode.
It blends newly drawn objects according to normal alpha blending, but
does not write the new alpha value into the destination buffer.  This is
useful in a couple spots, but.. my comments in the code are probably
right;  I should probably be exposing blend modes more directly instead
of adding new "Draw Modes".